Добавление страниц

+++ CodeShare_Library/Models/CodeSnippets.cs
@@ -15,10 +15,10 @@ namespace CodeShare_Library.Models
         public string Title { get; set; }
         public string Code { get; set; }
 
-        public int  UserId { get; set; }
+        public long UserId { get; set; }
         public string Description { get; set; }
 
-        public int IsAdmin { get; set; }
+        public long IsAdmin { get; set; }
 
         public string CreateAt { get; set; }
         public string UpdateAt { get; set; }
